+++ Services/ServicesContracts/TacheServiceC.cs
@@ -6,7 +6,7 @@ namespace Projet.Services.ServicesContracts
     {
         public Task AddTache(Tache tache);
         public Task<IEnumerable<Tache>> GetAllTache();
-        public Task<Tache> GetTacheByid(int id);
+        public Task<Tache> GetTacheByid(long id);
         public void UpdateTache(Tache tache);
         public Task DeleteTache(long id);
         public Task<IEnumerable<Tache>> GetByProjetIdAsync(int projetId);
